Fixed bug to JTaskPane color border and fixed son but in RootPane
I reintroduced the border color on the component JTaskPane and fixed the bug inside the life clycle.

Then, I introducing the bug fixing inside the MaterialRootComponent, now the windows was dragged in more monitor.

In conclusion, the bug inside the resizing exist yet.

This commit was built by @vincenzopalazzo vincenzopalazzodev@gmail.com